Fix QtWebEngine build on Windows
Work-around various issues to make Chromium build on MSVC or clang-cl

+++ chromium/v8/src/runtime/runtime-test.cc
@@ -1194,7 +1194,7 @@ RUNTIME_FUNCTION(Runtime_SetAllocationTimeout) {
 namespace {
 
 int FixedArrayLenFromSize(int size) {
-  return std::min({(size - OFFSET_OF_DATA_START(FixedArray)) / kTaggedSize,
+  return std::min({(int)(size - OFFSET_OF_DATA_START(FixedArray)) / kTaggedSize,
                    FixedArray::kMaxRegularLength});
 }
